requirements/requirements.txt
cassandra-driver==3.25.0
Django==4.0
django-cassandra-engine==1.6.3
django-crispy-forms==1.13.0
django-htmx==1.6.0